update openapi for immich 3.2

+++ internal/immich_open_api/immich_open_api.go
@@ -1424,7 +1424,7 @@ type AlbumResponseDto struct {
 	// Description Album description
 	Description string `json:"description"`
 
-	// EndDate End date (latest asset)
+	// EndDate UTC representation of (local) end date (latest asset)
 	EndDate *time.Time `json:"endDate,omitempty"`
 
 	// HasSharedLink Has shared link
@@ -1445,7 +1445,7 @@ type AlbumResponseDto struct {
 	// Shared Is shared album
 	Shared bool `json:"shared"`
 
-	// StartDate Start date (earliest asset)
+	// StartDate UTC representation of (local) start date (earliest asset)
 	StartDate *time.Time `json:"startDate,omitempty"`
 
 	// UpdatedAt Last update date
@@ -1611,10 +1611,10 @@ type AssetBulkUpdateDto struct {
 	IsFavorite *bool `json:"isFavorite,omitempty"`
 
 	// Latitude Latitude coordinate
-	Latitude *float32 `json:"latitude,omitempty"`
+	Latitude *float64 `json:"latitude,omitempty"`
 
 	// Longitude Longitude coordinate
-	Longitude *float32 `json:"longitude,omitempty"`
+	Longitude *float64 `json:"longitude,omitempty"`
 
 	// Rating Rating in range [1-5] (starred), -1 (rejected), or null (unrated)
 	Rating *int `json:"rating"`
@@ -2583,25 +2583,25 @@ type ExifResponseDto struct {
 	ExposureTime *string `json:"exposureTime"`
 
 	// FNumber F-number (aperture)
-	FNumber *float32 `json:"fNumber"`
+	FNumber *float64 `json:"fNumber"`
 
 	// FileSizeInByte File size in bytes
 	FileSizeInByte *int `json:"fileSizeInByte"`
 
 	// FocalLength Focal length in mm
-	FocalLength *float32 `json:"focalLength"`
+	FocalLength *float64 `json:"focalLength"`
 
 	// Iso ISO sensitivity
 	Iso *int `json:"iso"`
 
 	// Latitude GPS latitude
-	Latitude *float32 `json:"latitude"`
+	Latitude *float64 `json:"latitude"`
 
 	// LensModel Lens model
 	LensModel *string `json:"lensModel"`
 
 	// Longitude GPS longitude
-	Longitude *float32 `json:"longitude"`
+	Longitude *float64 `json:"longitude"`
 
 	// Make Camera make
 	Make *string `json:"make"`
@@ -3273,26 +3273,26 @@ type NotificationUpdateDto struct {
 
 // NumberFilter defines model for NumberFilter.
 type NumberFilter struct {
-	Eq    *float32   `json:"eq,omitempty"`
-	Gt    *float32   `json:"gt,omitempty"`
-	Gte   *float32   `json:"gte,omitempty"`
-	In    *[]float32 `json:"in,omitempty"`
-	Lt    *float32   `json:"lt,omitempty"`
-	Lte   *float32   `json:"lte,omitempty"`
-	Ne    *float32   `json:"ne,omitempty"`
-	NotIn *[]float32 `json:"notIn,omitempty"`
+	Eq    *float64   `json:"eq,omitempty"`
+	Gt    *float64   `json:"gt,omitempty"`
+	Gte   *float64   `json:"gte,omitempty"`
+	In    *[]float64 `json:"in,omitempty"`
+	Lt    *float64   `json:"lt,omitempty"`
+	Lte   *float64   `json:"lte,omitempty"`
+	Ne    *float64   `json:"ne,omitempty"`
+	NotIn *[]float64 `json:"notIn,omitempty"`
 }
 
 // NumberFilterNullable defines model for NumberFilterNullable.
 type NumberFilterNullable struct {
-	Eq    *float32   `json:"eq"`
-	Gt    *float32   `json:"gt,omitempty"`
-	Gte   *float32   `json:"gte,omitempty"`
-	In    *[]float32 `json:"in,omitempty"`
-	Lt    *float32   `json:"lt,omitempty"`
-	Lte   *float32   `json:"lte,omitempty"`
-	Ne    *float32   `json:"ne"`
-	NotIn *[]float32 `json:"notIn,omitempty"`
+	Eq    *float64   `json:"eq"`
+	Gt    *float64   `json:"gt,omitempty"`
+	Gte   *float64   `json:"gte,omitempty"`
+	In    *[]float64 `json:"in,omitempty"`
+	Lt    *float64   `json:"lt,omitempty"`
+	Lte   *float64   `json:"lte,omitempty"`
+	Ne    *float64   `json:"ne"`
+	NotIn *[]float64 `json:"notIn,omitempty"`
 }
 
 // OAuthAuthorizeResponseDto defines model for OAuthAuthorizeResponseDto.
@@ -3569,10 +3569,10 @@ type PlacesResponseDto struct {
 	Admin2name *string `json:"admin2name,omitempty"`
 
 	// Latitude Latitude coordinate
-	Latitude float32 `json:"latitude"`
+	Latitude float64 `json:"latitude"`
 
 	// Longitude Longitude coordinate
-	Longitude float32 `json:"longitude"`
+	Longitude float64 `json:"longitude"`
 
 	// Name Place name
 	Name string `json:"name"`
@@ -4005,7 +4005,7 @@ type ReverseGeocodingStateResponseDto struct {
 // RotateParameters defines model for RotateParameters.
 type RotateParameters struct {
 	// Angle Rotation angle in degrees
-	Angle float32 `json:"angle"`
+	Angle int `json:"angle"`
 }
 
 // SearchAlbumResponseDto defines model for SearchAlbumResponseDto.
@@ -5174,16 +5174,16 @@ type TimeBucketAssetResponseDto struct {
 	IsTrashed []bool `json:"isTrashed"`
 
 	// Latitude Array of latitude coordinates extracted from EXIF GPS data
-	Latitude *[]float32 `json:"latitude,omitempty"`
+	Latitude *[]float64 `json:"latitude,omitempty"`
 
 	// LivePhotoVideoId Array of live photo video asset IDs (null for non-live photos)
 	LivePhotoVideoId []string `json:"livePhotoVideoId"`
 
 	// LocalOffsetHours Array of UTC offset hours at the time each photo was taken. Positive values are east of UTC, negative values are west of UTC. Values may be fractional (e.g., 5.5 for +05:30, -9.75 for -09:45). Applying this offset to 'fileCreatedAt' will give you the time the photo was taken from the photographer's perspective.
-	LocalOffsetHours []float32 `json:"localOffsetHours"`
+	LocalOffsetHours []float64 `json:"localOffsetHours"`
 
 	// Longitude Array of longitude coordinates extracted from EXIF GPS data
-	Longitude *[]float32 `json:"longitude,omitempty"`
+	Longitude *[]float64 `json:"longitude,omitempty"`
 
 	// OwnerId Array of owner IDs for each asset
 	OwnerId []string `json:"ownerId"`
@@ -5192,7 +5192,7 @@ type TimeBucketAssetResponseDto struct {
 	ProjectionType []string `json:"projectionType"`
 
 	// Ratio Array of aspect ratios (width/height) for each asset
-	Ratio []float32 `json:"ratio"`
+	Ratio []float64 `json:"ratio"`
 
 	// Stack Array of stack information as [stackId, assetCount] tuples (null for non-stacked assets)
 	Stack *[][]string `json:"stack,omitempty"`
@@ -5264,13 +5264,13 @@ type UpdateAssetDto struct {
 	IsFavorite *bool `json:"isFavorite,omitempty"`
 
 	// Latitude Latitude coordinate
-	Latitude *float32 `json:"latitude,omitempty"`
+	Latitude *float64 `json:"latitude,omitempty"`
 
 	// LivePhotoVideoId Live photo video ID
 	LivePhotoVideoId *openapi_types.UUID `json:"livePhotoVideoId"`
 
 	// Longitude Longitude coordinate
-	Longitude *float32 `json:"longitude,omitempty"`
+	Longitude *float64 `json:"longitude,omitempty"`
 
 	// Rating Rating in range [1-5] (starred), -1 (rejected), or null (unrated)
 	Rating *int `json:"rating"`
